relation stablished between provider and product

+++ src/main/java/com/ecommerce/ecommerce/provider/ProviderController.java
@@ -27,7 +27,7 @@ public class ProviderController {
     }
 
     @GetMapping(path = "{id}")
-    public ResponseEntity<?> getProviderById(@PathVariable("id") Long id) {
+    public ResponseEntity<?> getProviderById(@PathVariable("id") Integer id) {
         Optional<Provider> providerOptional = providerService.getProviderById(id);
 
         if (providerOptional.isPresent()) {
