+++ src/main/java/com/ecommerce/ecommerce/provider/ProviderRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ProviderRepository extends JpaRepository<Provider, Long> {
+public interface ProviderRepository extends JpaRepository<Provider, Integer> {
 
     Optional<Provider> findByEmail(String email);
 }
